launchpad: Extract gas helpers from SigningCosmosClient file
