+++ packages/launchpad/src/index.ts
@@ -29,6 +29,7 @@ export {
   isSearchByTagsQuery,
 } from "./cosmosclient";
 export { makeSignBytes } from "./encoding";
+export { buildFeeTable, FeeTable, GasPrice } from "./gas";
 export {
   AuthAccountsResponse,
   AuthExtension,
@@ -97,7 +98,7 @@ export {
 } from "./pubkey";
 export { findSequenceForSignedTx } from "./sequence";
 export { encodeSecp256k1Signature, decodeSignature } from "./signature";
-export { FeeTable, SigningCosmosClient } from "./signingcosmosclient";
+export { SigningCosmosClient } from "./signingcosmosclient";
 export { isStdTx, pubkeyType, CosmosSdkTx, PubKey, StdFee, StdSignature, StdTx } from "./types";
 export {
   AccountData,
