+++ packages/launchpad/src/signingcosmosclient.spec.ts
@@ -3,9 +3,10 @@ import { assert } from "@cosmjs/utils";
 
 import { Coin, coin, coins } from "./coins";
 import { assertIsBroadcastTxSuccess, PrivateCosmosClient } from "./cosmosclient";
+import { GasPrice } from "./gas";
 import { MsgDelegate } from "./msgs";
 import { Secp256k1Wallet } from "./secp256k1wallet";
-import { GasPrice, SigningCosmosClient } from "./signingcosmosclient";
+import { SigningCosmosClient } from "./signingcosmosclient";
 import { makeRandomAddress, pendingWithoutWasmd, validatorAddress } from "./testutils.spec";
 
 const httpUrl = "http://localhost:1317";
